requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Mail==0.10.0
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
